pyproject.toml
[project]
name = "connectly"
version = "0.1.0"
description = ""
authors = [
    {name = "fkihai",email = "fikrihaikal568@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (==0.112.2)",
    "jwt (>=1.4.0,<2.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "sqlalchemy (>=2.0.43,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "pydantic-settings (>=2.10.1,<3.0.0)",
    "python-jose (>=3.5.0,<4.0.0)",
    "alembic (>=1.16.5,<2.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false

[tool.isort]
profile = "black"